requirements.txt
Django==2.2.24
xhtml2pdf>=0.2.3
gunicorn>=19.7.1
psycopg2==2.8.6
whitenoise>=5.0.1
djangorestframework>=3.12.4
#requests>=2.22.0
#django-cors-headers>=2.4.0